requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
Werkzeug==3.0.2
